requirements.txt
# Core ML packages
numpy
pandas
scikit-learn

# Database connection
psycopg2-binary

# Testing
pytest
